+++ src/form/Toggle/Toggle.tsx
@@ -1,5 +1,5 @@
 import React, { FC, forwardRef, LegacyRef } from 'react';
-import { motion } from 'motion/react';
+import { motion, MotionNodeAnimationOptions } from 'motion/react';
 import { ToggleSizeTheme, ToggleTheme } from './ToggleTheme';
 import { cn, useComponentTheme } from '@/utils';
 
@@ -13,6 +13,16 @@ export interface ToggleProps {
    * Whether the toggle is disabled or not.
    */
   disabled?: boolean;
+
+  /**
+   * Whether to animate the toggle.
+   */
+  animated?: boolean;
+
+  /**
+   * Animation configuration for the toggle.
+   */
+  animation?: MotionNodeAnimationOptions;
 
   /**
    * Additional class names to apply to the toggle.
@@ -59,6 +69,8 @@ export const Toggle: FC<ToggleProps & ToggleRef> = forwardRef<
       onBlur,
       className,
       size = 'medium',
+      animated = true,
+      animation,
       theme: customTheme,
       ...rest
     },
@@ -100,11 +112,16 @@ export const Toggle: FC<ToggleProps & ToggleRef> = forwardRef<
             [theme.handle.disabledAndChecked]: disabled && checked
           })}
           layout
-          transition={{
-            type: 'spring',
-            stiffness: 700,
-            damping: 30
-          }}
+          {...(animation
+            ? animation
+            : {
+                transition: {
+                  type: 'spring',
+                  stiffness: 700,
+                  damping: 30,
+                  duration: animated ? 0.3 : 0
+                }
+              })}
         />
       </div>
     );