+++ src/layers/Drawer/Drawer.story.tsx
@@ -86,10 +86,12 @@ export const CustomAnimation = () => {
     <Fragment>
       <Drawer
         className="max-w-[400px] mr-10 max-h-[90vh] my-auto rounded-xl"
-        initial={{ x: '120%', opacity: 0 }}
-        animate={{ x: 0, opacity: 1 }}
-        exit={{ x: '120%', opacity: 0 }}
-        transition={{ duration: 0.5, ease: [0.12, 0.78, 0.32, 0.98] }}
+        animation={{
+          initial: { x: '120%', opacity: 0 },
+          animate: { x: 0, opacity: 1 },
+          exit: { x: '120%', opacity: 0 },
+          transition: { duration: 0.5, ease: [0.12, 0.78, 0.32, 0.98] }
+        }}
         open={isOpen}
         onClose={() => setIsOpen(false)}
       >